requirements.txt
## The following requirements were added by pip freeze:
asttokens==2.4.1
black==24.4.2
blinker==1.8.2
branca==0.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
Flask-Mail==0.10.0
folium==0.16.0
fonttools==4.51.0
geographiclib==2.0
geopy==2.4.1
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kneed==0.8.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
MiniSom==2.3.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
requests==2.32.2
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
setuptools==71.0.3
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
xyzservices==2024.4.0
